+++ packages/accounts-controller/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Bump `@metamask/base-controller` from `^7.0.0` to `^7.1.0` ([#5079](https://github.com/MetaMask/core/pull/5079))
+- Rename `AccountsControllerMessenger` to `AccountsMessenger` ([#5090](https://github.com/MetaMask/core/pull/5090))
+  - The old name is preserved as an alias, so this is not a breaking change.
 
 ## [20.0.2]